+++ src/modules/components/proposal/proposalActions/proposalActionsItem/proposalActionsItemBasicView.test.tsx
@@ -9,7 +9,6 @@ jest.mock('../proposalActionsList', () => ({
     ProposalActionTokenMint: () => <div data-testid="token-mint-mock" />,
     ProposalActionUpdateMetadata: () => <div data-testid="update-metadata-mock" />,
     ProposalActionWithdrawToken: () => <div data-testid="withdraw-token-mock" />,
-    ProposalActionUpdatePluginMetadata: () => <div data-testid="update-plugin-metadata-mock" />,
 }));
 
 describe('<ProposalActionsItemBasicView /> component', () => {
@@ -43,7 +42,7 @@ describe('<ProposalActionsItemBasicView /> component', () => {
         { type: ProposalActionType.ADD_MEMBERS, testId: 'change-members-mock' },
         { type: ProposalActionType.REMOVE_MEMBERS, testId: 'change-members-mock' },
         { type: ProposalActionType.UPDATE_METADATA, testId: 'update-metadata-mock' },
-        { type: ProposalActionType.UPDATE_PLUGIN_METADATA, testId: 'update-plugin-metadata-mock' },
+        { type: ProposalActionType.UPDATE_PLUGIN_METADATA, testId: 'update-metadata-mock' },
     ])('renders $type component when action is of $type type', ({ type, testId }) => {
         const action = generateProposalAction({ type });
         render(createTestComponent({ action }));