+++ src/modules/components/proposal/proposalActions/proposalActionsItem/proposalActionsItemBasicView.tsx
@@ -8,7 +8,6 @@ import {
     ProposalActionChangeSettings,
     ProposalActionTokenMint,
     ProposalActionUpdateMetadata,
-    ProposalActionUpdatePluginMetadata,
     ProposalActionWithdrawToken,
 } from '../proposalActionsList';
 import { proposalActionsItemUtils } from './proposalActionsItemUtils';
@@ -50,10 +49,6 @@ export const ProposalActionsItemBasicView = <TAction extends IProposalAction = I
 
     if (proposalActionsItemUtils.isChangeSettingsAction(action)) {
         return <ProposalActionChangeSettings action={action} {...commonProps} />;
-    }
-
-    if (proposalActionsItemUtils.isUpdatePluginMetadataAction(action)) {
-        return <ProposalActionUpdatePluginMetadata action={action} {...commonProps} />;
     }
 
     return null;