+++ src/modules/components/proposal/proposalActions/proposalActionsList/index.ts
@@ -2,5 +2,4 @@ export * from './proposalActionChangeMembers';
 export * from './proposalActionChangeSettings';
 export * from './proposalActionTokenMint';
 export * from './proposalActionUpdateMetadata';
-export * from './proposalActionUpdatePluginMetadata';
 export * from './proposalActionWithdrawToken';