+++ src/modules/components/proposal/proposalActions/proposalActionsList/proposalActionUpdateMetadata/proposalActionUpdateMetadata.tsx
@@ -2,12 +2,18 @@ import { useState } from 'react';
 import { Avatar, DefinitionList, IconType, Link, Toggle, ToggleGroup } from '../../../../../../core';
 import { modulesCopy } from '../../../../../assets';
 import type { IProposalActionUpdateMetadataProps } from './proposalActionUpdateMetadata.api';
+import { ProposalActionType } from '../../proposalActionsDefinitions';
 
 export const ProposalActionUpdateMetadata: React.FC<IProposalActionUpdateMetadataProps> = (props) => {
     const [toggleValue, setToggleValue] = useState<string | undefined>('existingMetadata');
     const { action } = props;
     const { proposedMetadata, existingMetadata } = action;
     const metadataToDisplay = toggleValue === 'proposedMetadata' ? proposedMetadata : existingMetadata;
+
+    const daoMetadataAction = action.type === ProposalActionType.UPDATE_METADATA;
+    const pluginMetadataAction = action.type === ProposalActionType.UPDATE_PLUGIN_METADATA;
+
+    const descriptionTermLabel = daoMetadataAction ? 'descriptionTerm' : 'summaryTerm';
 
     return (
         <div className="flex w-full flex-col gap-2">
@@ -17,16 +23,23 @@ export const ProposalActionUpdateMetadata: React.FC<IProposalActionUpdateMetadat
             </ToggleGroup>
 
             <DefinitionList.Container>
-                <DefinitionList.Item
-                    className="md:items-center"
-                    term={modulesCopy.proposalActionsUpdateMetadata.logoTerm}
-                >
-                    <Avatar alt="dao-logo" src={metadataToDisplay.logo} responsiveSize={{ md: 'md', sm: 'sm' }} />
-                </DefinitionList.Item>
+                {daoMetadataAction && (
+                    <DefinitionList.Item
+                        className="md:items-center"
+                        term={modulesCopy.proposalActionsUpdateMetadata.logoTerm}
+                    >
+                        <Avatar alt="dao-logo" src={metadataToDisplay.logo} responsiveSize={{ md: 'md', sm: 'sm' }} />
+                    </DefinitionList.Item>
+                )}
                 <DefinitionList.Item term={modulesCopy.proposalActionsUpdateMetadata.nameTerm}>
                     <p className="text-base leading-tight text-neutral-800">{metadataToDisplay.name}</p>
                 </DefinitionList.Item>
-                <DefinitionList.Item term={modulesCopy.proposalActionsUpdateMetadata.descriptionTerm}>
+                {pluginMetadataAction && existingMetadata.key && (
+                    <DefinitionList.Item term={modulesCopy.proposalActionsUpdateMetadata.keyTerm}>
+                        <p className="text-base leading-tight text-neutral-800">{metadataToDisplay.key}</p>
+                    </DefinitionList.Item>
+                )}
+                <DefinitionList.Item term={modulesCopy.proposalActionsUpdateMetadata[descriptionTermLabel]}>
                     <p className="text-base leading-tight text-neutral-800">{metadataToDisplay.description}</p>
                 </DefinitionList.Item>
                 <DefinitionList.Item term={modulesCopy.proposalActionsUpdateMetadata.linkTerm}>