+++ src/modules/components/proposal/proposalActions/proposalActionsList/proposalActionUpdatePluginMetadata/index.ts
@@ -1,8 +1,0 @@
-export { ProposalActionUpdatePluginMetadata } from './proposalActionUpdatePluginMetadata';
-export type {
-    IProposalActionUpdatePluginMetadata,
-    IProposalActionUpdatePluginMetadataPluginMetadata,
-    IProposalActionUpdatePluginMetadataPluginMetadataLink,
-    IProposalActionUpdatePluginMetadataProps,
-} from './proposalActionUpdatePluginMetadata.api';
-export * from './proposalActionUpdatePluginMetadata.testUtils';